macos-port: native ARM64 baseline (D3D8->Metal shim + osdep_compat)
Initial drop of the macOS-native Apple-Silicon port of the engine. Boots,
loads, runs a stable 1v1 skirmish; LOW graphics preset is correct,
Medium/High still being ported (see MACOS_PORT_PLAN.md for status +
session-by-session technical history).

Architecture:
  * cmake/dx8_stub/        — D3D8 -> Metal shim (no Wine/DXVK/MoltenVK)
                             - metal_backend.{h,mm}: pipelines, FF lighting
                               + combiner emulated in MSL, depth, stencil,
                               BC1/2/3 textures, Cocoa input + cursor
                             - dx8_device.cpp: IDirect3DDevice8 impl
                             - gdi_text.mm: Core Text font surface backend
  * Dependencies/Utility/osdep_compat/  — Win32 API surface for POSIX/Cocoa
  * Cocoa input + AppKit window backend in Core/.../Win32Device/
  * Sparse #if defined(__APPLE__) blocks in engine sources for the few
    cases where engine-side awareness is unavoidable (LP64 fixups,
    sysctl-based CPU info, env-var-driven debug toggles).

Engine code stays as close to upstream as practical; macOS-specific
tweaks live behind #if defined(__APPLE__) so the Windows MSVC build
keeps compiling.

EA proprietary assets are NEVER committed — original/ is .gitignore'd.

See MACOS_PORT_PLAN.md for the deep technical narrative:
  - LP64 fixups (bittype.h, ddsfile.h LegacyDDSURFACEDESC2, TGA2Footer,
    persistfactory pointer/uint32 round-trip, DataChunk UTF-16 readers)
  - Texture pipeline (native BC1/2/3 on Apple Silicon, terrain BIG mounts)
  - Shroud TCI plumbing (D3DTSS_TCI_CAMERASPACEPOSITION end-to-end)
  - Cursor wiring (NSCursor + CGWarpMouseCursorPosition)
  - Stage 5 stencil (real MTLStencilDescriptor; killed the persistent
    player-color rect from RTS3DScene's flushOccludedObjectsIntoStencil)
  - Numerous null-deref guards and defensive clamps along the
    way (W3DWaypointBuffer, drawFillRect/drawOpenRect, font surface).

+++ Core/GameEngine/Source/GameClient/GUI/IMEManager.cpp
@@ -1431,7 +1431,7 @@ void IMEManager::updateCandidateList( Int candidateFlags  )
 
 						for( i=0; i < m_candidateCount; i++ )
 						{
-							Char *string = (Char*) ((UnsignedInt) clist + (UnsignedInt) clist->dwOffset[i]);
+							Char *string = (Char*) ((UnsignedInt)(uintptr_t) clist + (UnsignedInt) clist->dwOffset[i]);
 							if ( unicode )
 							{
 								m_candidateString[i].set( (WideChar *) string);
